requirements.txt
google-generativeai
python-dotenv
termcolor
argparse